Add TwilioHelperTests class
Added TwilioHelperTestClass, Moq dependency
