+++ tests/Adapters/Microsoft.Bot.Builder.Adapters.Twilio.Tests/TwilioAdapterTests.cs
@@ -2,12 +2,10 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Threading;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Bot.Schema;
-using Newtonsoft.Json;
+using Moq;
 using Xunit;
 
 namespace Microsoft.Bot.Builder.Adapters.Twilio.Tests
@@ -23,86 +21,58 @@ namespace Microsoft.Bot.Builder.Adapters.Twilio.Tests
         [Fact]
         public void Constructor_Should_Fail_With_Null_TwilioNumber()
         {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = null,
-                AccountSid = "Test",
-                AuthToken = "Test",
-            };
+            var options = new Mock<ITwilioAdapterOptions>();
+            options.SetupAllProperties();
+            options.Object.AuthToken = "Test";
+            options.Object.AccountSid = "Test";
 
-            Assert.Throws<Exception>(() => { new TwilioAdapter(options); });
+            Assert.Throws<Exception>(() => { new TwilioAdapter(options.Object); });
         }
 
         [Fact]
         public void Constructor_Should_Fail_With_Null_AccountSid()
         {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = "Test",
-                AccountSid = null,
-                AuthToken = "Test",
-            };
+            var options = new Mock<ITwilioAdapterOptions>();
+            options.SetupAllProperties();
+            options.Object.AuthToken = "Test";
+            options.Object.TwilioNumber = "Test";
 
-            Assert.Throws<Exception>(() => { new TwilioAdapter(options); });
+            Assert.Throws<Exception>(() => { new TwilioAdapter(options.Object); });
         }
 
         [Fact]
         public void Constructor_Should_Fail_With_Null_AuthToken()
         {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = "Test",
-                AccountSid = "Test",
-                AuthToken = null,
-            };
+            var options = new Mock<ITwilioAdapterOptions>();
+            options.SetupAllProperties();
+            options.Object.TwilioNumber = "Test";
+            options.Object.AccountSid = "Test";
 
-            Assert.Throws<Exception>(() => { new TwilioAdapter(options); });
+            Assert.Throws<Exception>(() => { new TwilioAdapter(options.Object); });
         }
 
         [Fact]
         public void Constructor_WithArguments_Succeeds()
         {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = "Test",
-                AccountSid = "Test",
-                AuthToken = "Test",
-            };
+            var options = new Mock<ITwilioAdapterOptions>();
+            options.SetupAllProperties();
+            options.Object.AuthToken = "Test";
+            options.Object.TwilioNumber = "Test";
+            options.Object.AccountSid = "Test";
 
-            Assert.NotNull(new TwilioAdapter(options));
-        }
-
-        [Fact]
-        public void ActivityToTwilio_Should_Return_MessageOptions_With_MediaUrl()
-        {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = "Test",
-                AccountSid = "Test",
-                AuthToken = "Test",
-            };
-
-            var adapter = new TwilioAdapter(options);
-            var activity = JsonConvert.DeserializeObject<Activity>(File.ReadAllText(Directory.GetCurrentDirectory() + @"\files\Activities.json"));
-            var activities = new Activity[] { activity };
-            activity.Attachments = new List<Attachment> { new Attachment(contentUrl: "http://example.com") };
-            var messageOption = TwilioHelper.ActivityToTwilio(activity, "123456789");
-
-            Assert.Equal(activity.Conversation.Id, messageOption.ApplicationSid);
-            Assert.Equal("123456789", messageOption.From.ToString());
-            Assert.Equal(activity.Text, messageOption.Body);
-            Assert.Equal(new Uri(activity.Attachments[0].ContentUrl), messageOption.MediaUrl[0]);
+            Assert.NotNull(new TwilioAdapter(options.Object));
         }
 
         [Fact]
         public async void SendActivitiesAsync_Should_Fail_With_ActivityType_Not_Message()
         {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = "Test", AccountSid = "Test", AuthToken = "Test",
-            };
+            var options = new Mock<ITwilioAdapterOptions>();
+            options.SetupAllProperties();
+            options.Object.AuthToken = "Test";
+            options.Object.TwilioNumber = "Test";
+            options.Object.AccountSid = "Test";
 
-            var twilioAdapter = new TwilioAdapter(options);
+            var twilioAdapter = new TwilioAdapter(options.Object);
 
             var activity = new Activity()
             {
@@ -120,12 +90,13 @@ namespace Microsoft.Bot.Builder.Adapters.Twilio.Tests
         [Fact]
         public async void ProcessAsync_Should_Fail_With_Null_HttpRequest()
         {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = "Test", AccountSid = "Test", AuthToken = "Test",
-            };
+            var options = new Mock<ITwilioAdapterOptions>();
+            options.SetupAllProperties();
+            options.Object.AuthToken = "Test";
+            options.Object.TwilioNumber = "Test";
+            options.Object.AccountSid = "Test";
 
-            var twilioAdapter = new TwilioAdapter(options);
+            var twilioAdapter = new TwilioAdapter(options.Object);
 
             await Assert.ThrowsAsync<ArgumentNullException>(async () =>
             {
@@ -136,12 +107,13 @@ namespace Microsoft.Bot.Builder.Adapters.Twilio.Tests
         [Fact]
         public async void ProcessAsync_Should_Fail_With_Null_HttpResponse()
         {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = "Test", AccountSid = "Test", AuthToken = "Test",
-            };
+            var options = new Mock<ITwilioAdapterOptions>();
+            options.SetupAllProperties();
+            options.Object.AuthToken = "Test";
+            options.Object.TwilioNumber = "Test";
+            options.Object.AccountSid = "Test";
 
-            var twilioAdapter = new TwilioAdapter(options);
+            var twilioAdapter = new TwilioAdapter(options.Object);
 
             await Assert.ThrowsAsync<ArgumentNullException>(async () =>
             {
@@ -152,28 +124,18 @@ namespace Microsoft.Bot.Builder.Adapters.Twilio.Tests
         [Fact]
         public async void ProcessAsync_Should_Fail_With_Null_Bot()
         {
-            ITwilioAdapterOptions options = new MockTwilioOptions
-            {
-                TwilioNumber = "Test", AccountSid = "Test", AuthToken = "Test",
-            };
+            var options = new Mock<ITwilioAdapterOptions>();
+            options.SetupAllProperties();
+            options.Object.AuthToken = "Test";
+            options.Object.TwilioNumber = "Test";
+            options.Object.AccountSid = "Test";
 
-            var twilioAdapter = new TwilioAdapter(options);
+            var twilioAdapter = new TwilioAdapter(options.Object);
 
             await Assert.ThrowsAsync<ArgumentNullException>(async () =>
             {
                 await twilioAdapter.ProcessAsync(default(HttpRequest), default(HttpResponse), null, default(CancellationToken));
             });
-        }
-
-        private class MockTwilioOptions : ITwilioAdapterOptions
-        {
-            public string TwilioNumber { get; set; }
-
-            public string AccountSid { get; set; }
-
-            public string AuthToken { get; set; }
-
-            public string ValidationUrl { get; set; }
         }
     }
 }
