+++ tests/Adapters/Microsoft.Bot.Builder.Adapters.Twilio.Tests/TwilioHelperTests.cs
@@ -1,0 +1,28 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using Microsoft.Bot.Schema;
+using Newtonsoft.Json;
+using Xunit;
+
+namespace Microsoft.Bot.Builder.Adapters.Twilio.Tests
+{
+    public class TwilioHelperTests
+    {
+        [Fact]
+        public void ActivityToTwilio_Should_Return_MessageOptions_With_MediaUrl()
+        {
+            var activity = JsonConvert.DeserializeObject<Activity>(File.ReadAllText(Directory.GetCurrentDirectory() + @"\files\Activities.json"));
+            activity.Attachments = new List<Attachment> { new Attachment(contentUrl: "http://example.com") };
+            var messageOption = TwilioHelper.ActivityToTwilio(activity, "123456789");
+
+            Assert.Equal(activity.Conversation.Id, messageOption.ApplicationSid);
+            Assert.Equal("123456789", messageOption.From.ToString());
+            Assert.Equal(activity.Text, messageOption.Body);
+            Assert.Equal(new Uri(activity.Attachments[0].ContentUrl), messageOption.MediaUrl[0]);
+        }
+    }
+}
